Make reset script idempotent for missing artifacts

+++ reset.sh
@@ -1,10 +1,14 @@
 #!/bin/sh
+
 rm -rf "work" "out"
-rm airootfs/root/packages/*.pkg.tar.zst
-rm airootfs/root/packages/*.pkg.tar.zst.sig
+rm -f airootfs/root/packages/*.pkg.tar.zst
+rm -f airootfs/root/packages/*.pkg.tar.zst.sig
 rm -rf airootfs/root/endeavouros-skel-liveuser/pkg
-rm airootfs/root/endeavouros-wallpaper.png
-rm airootfs/root/endeavouros-skel-liveuser/*.pkg.tar.zst
+rm -f airootfs/root/endeavouros-wallpaper.png
+rm -f airootfs/root/endeavouros-skel-liveuser/*.pkg.tar.zst
 rm -rf airootfs/etc/pacman.d/
-rm eosiso*.log
-mv airootfs/root/livewall-original.png airootfs/root/livewall.png
+rm -f eosiso*.log
+
+if [ -f airootfs/root/livewall-original.png ]; then
+	mv airootfs/root/livewall-original.png airootfs/root/livewall.png
+fi
